vim(plugins): remove 'not neovim' fallback
I've been using nvim for many years now. I almost never used "regular
vim" in this timeframe. I don't think this fallback is useful at all.
And when I did, it was clearly not in the context of a project where I'd
need a "git gutter" plugin, rather for quick config editing on servers
or such.

SO. Let's get rid of this alternate (and only) plugin.